osxbundle: add fonts.conf to distributed bundle
The osxbundle target creates a bundle that is supposed to be distributable
to third parties. As they may not have fontconfig installed they miss a
fonts.conf pointing to the usual fonts directories in OSX.

For people installing from source and using from the terminal this commit
changes nothing. You just have to make sure that your fontconfig is installed
with a sane configuration (XQuartz does). If you are installing fontconfig from
source you can force a sane OSX default using `--with-add-fonts`. For example:

  `./configure --with-add-fonts=/Library/Fonts,~/Library/Fonts`

Homebrew already addressed this with mxcl/homebrew@b242883